refactor: User 엔티티 경로 변화에 맞춰 import 경로 변경
