+++ src/main/java/trend_setter/turtlerun/inquiry/service/InquiryService.java
@@ -7,7 +7,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import trend_setter.turtlerun.inquiry.dto.InquiryListDto;
 import trend_setter.turtlerun.inquiry.dto.InquiryResponseDto;
-import trend_setter.turtlerun.inquiry.dto.InquirySearchDto;
 import trend_setter.turtlerun.inquiry.dto.InquiryStatusUpdateDto;
 import trend_setter.turtlerun.inquiry.dto.InquiryWriteDto;
 import trend_setter.turtlerun.inquiry.dto.InquiryDetailDto;
@@ -17,7 +16,7 @@ import trend_setter.turtlerun.inquiry.repository.InquiryResponseRepository;
 import trend_setter.turtlerun.inquiry.utils.InquiryMapper;
 import trend_setter.turtlerun.inquiry.repository.InquiryRepository;
 import trend_setter.turtlerun.inquiry.utils.InquiryResponseMapper;
-import trend_setter.turtlerun.user.User;
+import trend_setter.turtlerun.user.entity.User;
 import trend_setter.turtlerun.user.repository.UserRepository;
 
 @Service
